Replace the homepage diagram with the product page one
