+++ src/components/DiagramConnectors/types.ts
@@ -1,0 +1,3 @@
+export interface DiagramConnectorsProps {
+    hasTooltips?: boolean;
+}
