+++ src/components/Homepage/SectionTwo/index.tsx
@@ -1,31 +1,12 @@
-import { useMediaQuery, useTheme } from '@mui/material';
 import { StaticImage } from 'gatsby-plugin-image';
 import * as React from 'react';
+
+import DiagramSourceConnectors from '../../DiagramConnectors/DiagramSourceConnectors';
+import DiagramDestinationConnectors from '../../DiagramConnectors/DiagramDestinationConnectors';
 import {
-    AmazonKinesisLogoWrapper,
-    AmazonRedshiftLogoWrapper,
-    AmazonS3LogoWrapper,
-    ApacheKafkaLogoWrapper,
-    AzureBlobStorageLogoWrapper,
     Container,
-    DatabricksLogoWrapper,
-    DynamoDBLogoWrapper,
-    ElasticLogoWrapper,
-    FlowImageWrapper,
-    GoogleBigQueryLogoWrapper,
-    GoogleStorageLogoWrapper,
+    DesktopDiagramWrapper,
     Header,
-    HttpWebhookLogoWrapper,
-    HubspotLogoWrapper,
-    MongoDBLogoWrapper,
-    MySQLLogoWrapper,
-    NetsuiteLogoWrapper,
-    OpenAiLogoWrapper,
-    OracleLogoWrapper,
-    PineconeLogoWrapper,
-    PostgresLogoWrapper,
-    SalesforceLogoWrapper,
-    SnowflakeLogoWrapper,
     Step,
     StepDescription,
     StepIcon,
@@ -39,9 +20,6 @@ import {
 } from './styles';
 
 const SectionTwo = () => {
-    const theme = useTheme();
-    const isMobile = useMediaQuery(theme.breakpoints.down(768));
-
     return (
         <Container>
             <Header>
@@ -110,109 +88,25 @@ const SectionTwo = () => {
                     </div>
                 </Step>
             </Steps>
-            <FlowImageWrapper>
-                {isMobile ? (
-                    <StaticImage
-                        placeholder="none"
-                        alt="Estuary Flow streamlines complex data workflows by integrating diverse data sources and destinations"
-                        src="../../../images/flow-mobile.png"
-                        layout="constrained"
-                        quality={100}
-                    />
-                ) : (
-                    <StaticImage
-                        placeholder="none"
-                        alt="Estuary Flow streamlines complex data workflows by integrating diverse data sources and destinations"
-                        src="../../../images/flow-desktop.png"
-                        layout="constrained"
-                        quality={100}
-                    />
-                )}
-                <OracleLogoWrapper
-                    to="/source/oracle-flashback"
-                    aria-label="Oracle source"
+            <DesktopDiagramWrapper className="desktop-image">
+                <DiagramSourceConnectors />
+                <StaticImage
+                    placeholder="none"
+                    alt="Estuary Flow streamlines complex data workflows by integrating diverse data sources and destinations"
+                    src="../../../images/flow-cycle.png"
+                    layout="constrained"
+                    quality={100}
                 />
-                <MySQLLogoWrapper
-                    to="/source/mysql"
-                    aria-label="MySQL source"
-                />
-                <PostgresLogoWrapper
-                    to="/source/postgres"
-                    aria-label="Postgres source"
-                />
-                <AmazonS3LogoWrapper
-                    to="/source/s3"
-                    aria-label="Amazon S3 source"
-                />
-                <GoogleStorageLogoWrapper
-                    to="/source/gcs"
-                    aria-label="Google Storage source"
-                />
-                <AzureBlobStorageLogoWrapper
-                    to="/source/azure-blob-storage"
-                    aria-label="Azure source"
-                />
-                <NetsuiteLogoWrapper
-                    to="/source/netsuite-v"
-                    aria-label="Netsuite source"
-                />
-                <HubspotLogoWrapper
-                    to="/source/hubspot"
-                    aria-label="HubSpot source"
-                />
-                <SalesforceLogoWrapper
-                    to="/source/salesforce"
-                    aria-label="Salesforce source"
-                />
-                <HttpWebhookLogoWrapper
-                    to="/source/http-ingest"
-                    aria-label="Webhook source"
-                />
-                <AmazonKinesisLogoWrapper
-                    to="/source/kinesis"
-                    aria-label="Amazon Kinesis source"
-                />
-                <ApacheKafkaLogoWrapper
-                    to="/source/kafka"
-                    aria-label="Apache Kafka source"
-                />
-                <SnowflakeLogoWrapper
-                    to="/destination/snowflake"
-                    aria-label="Snowflake destination"
-                />
-                <GoogleBigQueryLogoWrapper
-                    to="/destination/bigquery"
-                    aria-label="Google BigQuery destination"
-                />
-                <AmazonRedshiftLogoWrapper
-                    to="/destination/redshift"
-                    aria-label="Amazon Redshift destination"
-                />
-                <ElasticLogoWrapper
-                    to="/destination/elasticsearch"
-                    aria-label="Elastic destination"
-                />
-                <MongoDBLogoWrapper
-                    to="/destination/mongodb"
-                    aria-label="MongoDB destination"
-                />
-                <DynamoDBLogoWrapper
-                    to="/destination/dynamodb"
-                    aria-label="DynamoDB destination"
-                />
-                <PineconeLogoWrapper
-                    to="/destination/pinecone"
-                    aria-label="Pinecone destination"
-                />
-                <OpenAiLogoWrapper
-                    to="/destination/pinecone"
-                    aria-label="Pinecone destination"
-                />
-                <DatabricksLogoWrapper
-                    to="/destination/databricks"
-                    aria-label="Databricks destination"
-                />
-            </FlowImageWrapper>
+                <DiagramDestinationConnectors />
+            </DesktopDiagramWrapper>
+            <StaticImage
+                placeholder="none"
+                alt="Estuary Flow streamlines complex data workflows by integrating diverse data sources and destinations"
+                src="../../../images/flow-mobile.png"
+                layout="constrained"
+                quality={100}
+                className="mobile-image"
+            />
         </Container>
     );
 };
